0825/danny/app/requirements.txt
python-dotenv
flask[async]
flask_sqlalchemy
Flask-WTF
email-validator